restructured project to be installable module
