+++ README.md
@@ -1,0 +1,12 @@
+# Breakaway Dice Mapper
+A dice probability sim for the development of [hit Sci-Fi Tabletop RPG Breakaway](http://www.breakaway.pub/)
+
+## Installation
+1. Git clone
+2. Enter Breakaway-Dice-Mapper directory
+3. pip install .
+
+## Dev Installation
+1. Git clone
+2. Enter Breakaway-Dice-Mapper directory
+3. pip install -e .
