--- output/5x5_d6_legacy_map.csv
@@ -1,10 +1,0 @@
-				DM Rolls					
-		1	2	3	4	5	6	7	8
-	1	0.5833333333333334	0.4444444444444444	0.375	0.3333333333333333	0.3055555555555556	0.2857142857142857	0.2708333333333333	0.25925925925925924
-	2	0.7222222222222222	0.7619047619047619	0.6428571428571429	0.5714285714285714	0.5238095238095238	0.4897959183673469	0.4642857142857143	0.4444444444444444
-	3	0.7916666666666666	0.8452380952380952	0.6294642857142857	0.5076530612244898	0.4302721088435374	0.37712585034013607	0.3385416666666667	0.30934343434343436
-	4	0.8333333333333334	0.8888888888888888	0.7304421768707483	0.7320483749055178	0.62600781053162	0.5525321239606954	0.49880751964085296	0.45791245791245794
-Player Rolls	5	0.8611111111111112	0.9149659863945578	0.7916666666666666	0.8053665910808768	0.6526832955404384	0.5500669964955679	0.4769871332371332	0.4226236726236726
-	6	0.8809523809523809	0.9319727891156463	0.8320578231292517	0.8510959939531368	0.7314900707757851	0.7236699087348438	0.6324965018146836	0.5641547005183368
-	7	0.8958333333333334	0.9437830687830688	0.8603670634920635	0.8816237774571107	0.7846971701138368	0.7859110586383313	0.66678145087236	0.5790843555995071
-	8	0.9074074074074074	0.9523809523809523	0.8811327561327561	0.903078403078403	0.8224460724460725	0.8287217832672378	0.7313303979970647	0.720704746345772
