--- output/5x5_d6_map.csv
@@ -1,7 +1,0 @@
-		DM Rolls				
-		1	2	3	4	5
-	1	0.5833333333333334	0.2777777777777778	0.20833333333333334	0.16666666666666666	0.1388888888888889
-	2	0.7222222222222222	0.6031746031746031	0.4166666666666667	0.3333333333333333	0.2777777777777778
-Player Rolls	3	0.7916666666666666	0.8452380952380952	0.5625	0.2695578231292517	0.20833333333333334
-	4	0.8333333333333334	0.8888888888888888	0.7304421768707483	0.5563114134542706	0.35257621567145375
-	5	0.8611111111111112	0.9149659863945578	0.7916666666666666	0.8053665910808768	0.5440287226001512
