--- output/8x8_d6_map.csv
@@ -1,10 +1,0 @@
-				DM Rolls					
-		1	2	3	4	5	6	7	8
-	1	0.5833333333333334	0.2777777777777778	0.20833333333333334	0.16666666666666666	0.1388888888888889	0.11904761904761904	0.10416666666666667	0.09259259259259259
-	2	0.7222222222222222	0.6031746031746031	0.4166666666666667	0.3333333333333333	0.2777777777777778	0.23809523809523808	0.20833333333333334	0.18518518518518517
-	3	0.7916666666666666	0.8452380952380952	0.5625	0.2695578231292517	0.20833333333333334	0.1679421768707483	0.1396329365079365	0.11886724386724387
-	4	0.8333333333333334	0.8888888888888888	0.7304421768707483	0.5563114134542706	0.35257621567145375	0.2857142857142857	0.23855619688953023	0.2037838704505371
-Player Rolls	5	0.8611111111111112	0.9149659863945578	0.7916666666666666	0.8053665910808768	0.5440287226001512	0.26850992922421496	0.21530282988616323	0.17755392755392754
-	6	0.8809523809523809	0.9319727891156463	0.8320578231292517	0.8510959939531368	0.7314900707757851	0.5393686400179907	0.3277936289299926	0.27177368086459
-	7	0.8958333333333334	0.9437830687830688	0.8603670634920635	0.8816237774571107	0.7846971701138368	0.7859110586383313	0.5335807570656056	0.2686696020029353
-	8	0.9074074074074074	0.9523809523809523	0.8811327561327561	0.903078403078403	0.8224460724460725	0.8287217832672378	0.7313303979970647	0.5304440013064722
