--- output/test.csv
@@ -1,8 +1,0 @@
-			DM Rolls				
-		1	2	3	4	5	6
-	1	0.1	0.1	0.1	0.1	0.1	0.1
-	2	0.1	0.1	0.1	0.1	0.1	0.1
-	3	0.1	0.1	0.1	0.1	0.1	0.1
-Player Rolls	4	0.1	0.1	0.1	0.1	0.1	0.1
-	5	0.1	0.1	0.1	0.1	0.1	0.1
-	6	0.1	0.1	0.1	0.1	0.1	0.1
